% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_find.R
\name{file_find}
\alias{file_find}
\title{Find Things in files}
\usage{
file_find(x, pattern, open = TRUE)
}
\arguments{
\item{x}{a character vector of files. I like \code{fs::dir_ls()}.}

\item{pattern}{A pattern to match.}

\item{open}{Logical indicator.}
}
\value{
Opens files at searched locations
}
\description{
Find Things in files
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
file_find("README.md", "install.packages")
\dontshow{\}) # examplesIf}
}
