% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{GetResultRanges}
\alias{GetResultRanges}
\title{This function returns the range of AUC scores and Monotonicity scores.
It also saves to a file if specified.}
\usage{
GetResultRanges(resultList)
}
\arguments{
\item{resultList}{A list of FERRET_ROC_AUC objects over which to compute ranges.}
}
\value{
A data frame with the following format:
Score, Statistic, Jaccard, Ingroup, Outgroup
RAUC, min, , ,
RAUC, max, , ,
RAUC, average, , ,
Monotonicity, min, , ,
Monotonicity, max, , ,
Monotonicity, average, , ,
}
\description{
This function returns the range of AUC scores and Monotonicity scores.
It also saves to a file if specified.
}
