% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{MakePerformanceBarPlot}
\alias{MakePerformanceBarPlot}
\title{This function generates separate bar plot of the performance for each cell
type across validation methods.}
\usage{
MakePerformanceBarPlot(resultDf, metric)
}
\arguments{
\item{resultDf}{A data frame with the following format:
Method, CellType, Score, Statistic, Jaccard, Ingroup, Outgroup}

\item{metric}{Metric used for network similarity}
}
\description{
This function generates separate bar plot of the performance for each cell
type across validation methods.
}
