% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{RunPathwayAnalysis}
\alias{RunPathwayAnalysis}
\title{This function runs pathway analysis for the network targets.}
\usage{
RunPathwayAnalysis(network, pathways, geneSymbols)
}
\arguments{
\item{network}{The input network, formatted as a matrix.}

\item{pathways}{A pathway object, generated using gmtPathways() from
the fgsea R package. The format of this object is a named list of 
vectors, where the elements of each vector are gene symbols.}

\item{geneSymbols}{The full list of gene symbols, with possible duplicates.}
}
\value{
The pathway results formatted as a data frame.
}
\description{
This function runs pathway analysis for the network targets.
}
